Move /components to /view & alias the path
