+++ client/src/views/dashboard/Registration.vue
